[scheduler][primitive] Move  and all related types to the  folder
