+++ packages/x-scheduler/package.json
@@ -27,6 +27,7 @@
     "./primitives/time-grid": "./src/primitives/time-grid/index.ts",
     "./primitives/timeline": "./src/primitives/timeline/index.ts",
     "./primitives/use-day-list": "./src/primitives/use-day-list/index.ts",
+    "./primitives/use-event-calendar": "./src/primitives/use-event-calendar/index.ts",
     "./primitives/use-week-list": "./src/primitives/use-week-list/index.ts"
   },
   "sideEffects": false,
