+++ packages/x-scheduler/src/material/internals/components/event-popover/EventPopover.tsx
@@ -16,8 +16,8 @@ import {
 import { useAdapter } from '../../../../primitives/utils/adapter/useAdapter';
 import { getColorClassName } from '../../utils/color-utils';
 import { useTranslations } from '../../utils/TranslationsContext';
-import { CalendarEvent } from '../../../models/events';
-import { selectors } from '../../../event-calendar/store';
+import { CalendarEvent } from '../../../../primitives/models';
+import { selectors } from '../../../../primitives/use-event-calendar';
 import { useEventCalendarContext } from '../../hooks/useEventCalendarContext';
 import { useSelector } from '../../../../base-ui-copy/utils/store';
 import { useEventCallback } from '../../../../base-ui-copy/utils/useEventCallback';
