+++ packages/x-scheduler/src/material/internals/components/event/day-grid-event/DayGridEvent.tsx
@@ -8,7 +8,7 @@ import { DayGridEventProps } from './DayGridEvent.types';
 import { getColorClassName } from '../../../utils/color-utils';
 import { useTranslations } from '../../../utils/TranslationsContext';
 import { useSelector } from '../../../../../base-ui-copy/utils/store';
-import { selectors } from '../../../../event-calendar/store';
+import { selectors } from '../../../../../primitives/use-event-calendar';
 import { useEventCalendarContext } from '../../../hooks/useEventCalendarContext';
 import './DayGridEvent.css';
 import '../index.css';
