+++ packages/x-scheduler/src/material/internals/components/event/time-grid-event/TimeGridEvent.tsx
@@ -7,7 +7,7 @@ import { useSelector } from '../../../../../base-ui-copy/utils/store';
 import { getAdapter } from '../../../../../primitives/utils/adapter/getAdapter';
 import { TimeGrid } from '../../../../../primitives/time-grid';
 import { getColorClassName } from '../../../utils/color-utils';
-import { selectors } from '../../../../event-calendar/store';
+import { selectors } from '../../../../../primitives/use-event-calendar';
 import { useEventCalendarContext } from '../../../hooks/useEventCalendarContext';
 import './TimeGridEvent.css';
 import '../index.css';
