+++ packages/x-scheduler/src/material/internals/components/header-toolbar/view-switcher/ViewSwitcher.tsx
@@ -5,11 +5,11 @@ import { Menu } from '@base-ui-components/react/menu';
 import { useForkRef } from '@base-ui-components/react/utils';
 import { ChevronDown } from 'lucide-react';
 import { Menubar } from '@base-ui-components/react/menubar';
-import { EventCalendarView } from '../../../../event-calendar/EventCalendar.types';
+import { CalendarView } from '../../../../../primitives/models';
 import { useTranslations } from '../../../utils/TranslationsContext';
 import { useEventCalendarContext } from '../../../hooks/useEventCalendarContext';
 import { useSelector } from '../../../../../base-ui-copy/utils/store';
-import { selectors } from '../../../../event-calendar/store';
+import { selectors } from '../../../../../primitives/use-event-calendar';
 import './ViewSwitcher.css';
 
 export const ViewSwitcher = React.forwardRef(function ViewSwitcher(
@@ -30,7 +30,7 @@ export const ViewSwitcher = React.forwardRef(function ViewSwitcher(
     (event: React.MouseEvent<HTMLElement>) => {
       const newView = event.currentTarget.getAttribute('data-view');
       if (newView) {
-        instance.setView(newView as EventCalendarView, event);
+        instance.setView(newView as CalendarView, event);
       }
     },
     [instance],
