+++ packages/x-scheduler/src/material/models/translations.ts
@@ -1,4 +1,4 @@
-import { EventCalendarView } from '../event-calendar/EventCalendar.types';
+import { CalendarView } from '../../primitives/models';
 
 export interface SchedulerTranslations {
   // ResourceLegend
@@ -19,9 +19,9 @@ export interface SchedulerTranslations {
 
   // MonthView
   hiddenEvents: (hiddenEventsCount: number) => string;
-  nextTimeSpan: (timeSpan: EventCalendarView) => string;
+  nextTimeSpan: (view: CalendarView) => string;
   noResourceAriaLabel: string;
-  previousTimeSpan: (timeSpan: EventCalendarView) => string;
+  previousTimeSpan: (view: CalendarView) => string;
   resourceAriaLabel: (resourceName: string) => string;
   weekAbbreviation: string;
   weekNumberAriaLabel: (weekNumber: number) => string;
