+++ packages/x-scheduler/src/material/month-view/MonthView.test.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { DateTime } from 'luxon';
 import { createSchedulerRenderer } from 'test/utils/scheduler';
 import { screen, within } from '@mui/internal-test-utils';
-import { CalendarEvent } from '@mui/x-scheduler/material';
+import { CalendarEvent } from '@mui/x-scheduler/primitives/models';
 import { spy } from 'sinon';
 import { MonthView } from '@mui/x-scheduler/material/month-view';
 import { StandaloneView } from '@mui/x-scheduler/material/standalone-view';
