+++ packages/x-scheduler/src/primitives/models/event.ts
@@ -1,0 +1,46 @@
+import type { SchedulerValidDate } from './date';
+import type { CalendarResourceId } from './resource';
+
+export interface CalendarEvent {
+  /**
+   * The unique identifier of the event.
+   */
+  id: CalendarEventId;
+  /**
+   * The title of the event.
+   */
+  title: string;
+  /**
+   * The description of the event.
+   */
+  description?: string;
+  /**
+   * The start date and time of the event.
+   */
+  start: SchedulerValidDate;
+  /**
+   * The end date and time of the event.
+   */
+  end: SchedulerValidDate;
+  /**
+   * The id of the resource this event is associated with.
+   */
+  resource?: CalendarResourceId;
+}
+
+export type CalendarEventId = string | number;
+
+// TODO: Try to keep this property on the Material UI side
+export type CalendarEventColor =
+  | 'primary'
+  | 'mauve'
+  | 'violet'
+  | 'cyan'
+  | 'jade'
+  | 'red'
+  | 'lime'
+  | 'orange'
+  | 'yellow'
+  | 'pink'
+  | 'indigo'
+  | 'blue';
