+++ packages/x-scheduler/src/primitives/models/view.ts
@@ -1,0 +1,1 @@
+export type CalendarView = 'week' | 'day' | 'month' | 'agenda';
