+++ packages/x-scheduler/src/primitives/use-event-calendar/index.ts
@@ -1,0 +1,3 @@
+export * from './useEventCalendar';
+export { selectors } from './store';
+export type { EventCalendarStore } from './store';
